Remove AbstractListParams.php
